requirements.txt
openenv-core[core]>=0.2.2
pydantic>=2.0.0
uvicorn>=0.20.0
openai>=1.0.0
